+++ mail/mail-uniffi/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
   - Fix failed conversation updates during draft save.
+  - [ET-2039] Scroller now reports an error when there is no more cached data and the api responds with an error.
 
 ## [0.40.4] - 2025-01-31